Separate rendering from game logic and change dimensioning calculation
- Create new `BoardRenderer` struct with atributtes and methods extracted from `Game`.
- Window height and width are calculated from `NROWS` and `NCOLS` in main.rs.
- Update docs and lib.rs.
